requirements.txt
click>=8.1.7
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
flask>=3.0.0
